Update vehicle data handling: change speed to double, remove resetTrip method duplication, and include CANReader header

+++ qt_app/src/vehicledata.cpp
@@ -142,7 +142,7 @@ void VehicleData::changeGearDown()
     }
 }
 
-int VehicleData::getSpeed() const
+double VehicleData::getSpeed() const
 {
 	return m_speed;
 }
